+++ packages/pwc/src/utils/checkTypes.ts
@@ -1,4 +1,5 @@
-import { TemplateFlag } from '../constants';
+import { TemplateData, TemplateFlag, TemplateString } from '../constants';
+import { hasOwnProperty } from './common';
 
 export const EMPTY_OBJECT = {};
 
@@ -35,7 +36,7 @@ export function isPrivate(name: string) {
 }
 
 export function isTemplate(value: unknown): boolean {
-  return value && value[TemplateFlag] === true;
+  return value && value[TemplateFlag] === true && hasOwnProperty(value, TemplateString) && hasOwnProperty(value, TemplateData);
 }
 
 export function isFalsy(value: unknown) {
